Added a simple spawning system for baits
Added a system that spawn baits on a semi-random timer. Bait now spawns, targets a set number of random tiles and then exits the screen.

Still need to improve movement as it is choppy and all harsh angles. Can now move on to create individual baits and adding their unique effects.