Added username setting
Added username option to the menu that allows users to change name that appears on leaderboard.

Also added the unit test for the API that I totally 100% definitely made before I started coding, it was just in the git ignore or something, a glitch, or maybe my ping was too high? Not exactly sure but its here now.